pyproject.toml
[project]
name = "gamma"
version = "0.1.0"
description = "Gamma assistant scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "sqlmodel>=0.0.22",
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
  "faster-whisper>=1.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["gamma*"]
